윤새영(대댓글완성)

+++ Project_1_Busan/src/main/java/com/sist/dao/ReplyDAO.java
@@ -28,10 +28,11 @@ public class ReplyDAO {
 		   try
 		   {
 			   conn=dbconn.getConnection();
-			   String sql="SELECT rrno,rno,id,cont,"
-			   		     +"TO_CHAR(regdate,'YYYY-MM-DD'),group_id,group_tab "
-					     +"FROM review_reply WHERE rno=?"
-					     + "ORDER BY group_id DESC,group_step ASC";
+			   String sql = "SELECT rrno, rno, id, cont, " +
+			             "TO_CHAR(regdate, 'YYYY-MM-DD \"(\"HH24:MI\")\"'), group_id, group_tab " +
+			             "FROM review_reply WHERE rno=? " +
+			             "ORDER BY group_id DESC, group_step ASC";
+
 			   		     
 			   ps=conn.prepareStatement(sql);
 			   ps.setInt(1, rno);
@@ -86,16 +87,42 @@ public class ReplyDAO {
 	   {
 		   try
 		   {
+			   
 			   conn=dbconn.getConnection();
-			   String sql="INSERT INTO review_reply(rrno,rno,id,cont,group_id,group_step,group_tab,root) "
-			   		+ "VALUES(rr_rrno_seq.nextval,?,?,?,?,?,?,?)";
+			   String sql="SELECT group_step,group_tab FROM review_reply WHERE rrno="+vo.getRoot();
+			   ps=conn.prepareStatement(sql);
+			   ResultSet rs=ps.executeQuery();
+			   rs.next();
+			   int groupStep=rs.getInt(1);
+			   int groupTab=rs.getInt(2);
+			   rs.close();
+			   ps.close();
+			   
+
+			   sql="UPDATE review_reply SET group_step=group_step+1 WHERE group_id="+vo.getGroup_id()+" AND group_step>"+groupStep;
+			   ps=conn.prepareStatement(sql);
+			   ps.executeUpdate();
+			   ps.close();
+			   
+			   
+			   sql="INSERT INTO review_reply(rrno,rno,id,cont,group_id,group_step,group_tab,root) "
+			   		+ "VALUES((SELECT NVL(MAX(rrno)+1,1) FROM review_reply),?,?,?,?,?,?,?)";
 			   		     
 			   ps=conn.prepareStatement(sql);
 			   ps.setInt(1, vo.getRno());
 			   ps.setString(2, vo.getId());
 			   ps.setString(3, vo.getCont());
 			   ps.setInt(4, vo.getGroup_id());
+			   ps.setInt(5, groupStep+1);
+			   ps.setInt(6, groupTab+1);
+			   ps.setInt(7, vo.getRoot());
 			   ps.executeUpdate();
+			   ps.close();
+			   
+			   sql="UPDATE review_reply SET depth=depth+1 WHERE rrno="+vo.getRoot();
+			   ps=conn.prepareStatement(sql);
+			   ps.executeUpdate();
+			   ps.close();
 			   
 		   }catch(Exception ex) 
 		   {
@@ -112,11 +139,12 @@ public class ReplyDAO {
 		   {
 			   conn=dbconn.getConnection();
 			   String sql = "INSERT INTO review_reply(rrno, rno, id, cont, group_id, group_step, group_tab) " +
-			             "VALUES(rr_rrno_seq.nextval, ?, ?, ?, (SELECT NVL(MAX(group_id) + 1, 1) FROM review_reply), 0, 0)";
+			             "VALUES((SELECT NVL(MAX(rrno)+1,1) FROM review_reply), ?, ?, ?, (SELECT NVL(MAX(group_id) + 1, 1) FROM review_reply WHERE rno=?), 0, 0)";
 			ps = conn.prepareStatement(sql);
 			ps.setInt(1, vo.getRno());
 			ps.setString(2, vo.getId());
 			ps.setString(3, vo.getCont());
+			ps.setInt(4,vo.getRno());
 			ps.executeUpdate();
 
 			   
@@ -134,10 +162,40 @@ public class ReplyDAO {
 		  try
 		   {
 			   conn=dbconn.getConnection();
-			   String sql = "DELETE FROM review_reply WHERE rrno=?";
-			ps = conn.prepareStatement(sql);
-			ps.setInt(1, rrno);
-			ps.executeUpdate();
+			   String sql = "SELECT root,depth FROM review_reply WHERE rrno="+rrno;
+			   ps=conn.prepareStatement(sql);
+			   ResultSet rs=ps.executeQuery();
+			   rs.next();
+			   int db_root=rs.getInt(1);
+			   int db_depth=rs.getInt(2);
+			   rs.close();
+			   ps.close();
+			   
+			   if(db_depth>0)
+			   {
+				   sql="UPDATE review_reply SET cont=? WHERE rrno="+rrno;
+				   ps=conn.prepareStatement(sql);
+				   ps.setString(1, "대댓글이 있는 경우 삭제할 수 없습니다.");
+				   ps.executeUpdate();
+				   ps.close();
+			   }
+			   else {
+				    sql = "DELETE FROM review_reply WHERE rrno=?";
+					ps = conn.prepareStatement(sql);
+					ps.setInt(1, rrno);
+					ps.executeUpdate();
+					ps.close();
+					
+					
+					if(db_root!=0)
+					{
+						sql="UPDATE review_reply SET depth=depth-1 WHERE rrno="+db_root;
+						ps = conn.prepareStatement(sql);
+						ps.executeUpdate();
+						ps.close();
+					}
+			   }
+			   
 			   
 		   }catch(Exception ex) 
 		   {
